Move carreirs API files to carriers module
